Replace Image component references with ImgixImage component
