+++ packages/components/src/Image/index.ts
@@ -1,2 +1,3 @@
-export { default as FastlyImage } from './FastlyImage';
 export * from './Image';
+export { default as Image } from './Image';
+export * from './ImgixImage';
