+++ packages/components/src/LegacyHeader/components/LegacyHeaderCampaign.d.ts
@@ -1,0 +1,1 @@
+export declare const HeaderCampaign: import("styled-components").StyledComponent<"div", any, import("../../Box").BoxProps, never>;
