+++ packages/components/src/LegacyHeader/components/LegacyHeaderCampaign.tsx
@@ -1,0 +1,12 @@
+import styled from 'styled-components';
+
+import { Box } from '../..';
+
+// eslint-disable-next-line import/prefer-default-export
+export const HeaderCampaign = styled(Box)`
+  overflow: hidden;
+
+  > a {
+    display: flex;
+  }
+`;
