+++ packages/components/src/LegacyHeader/components/LegacyHeaderLink.d.ts
@@ -1,0 +1,1 @@
+export declare const HeaderLink: import("styled-components").StyledComponent<"a", any, {}, never>;
