+++ packages/components/src/LegacyHeader/components/LegacyMainNav.tsx
@@ -14,7 +14,7 @@ import { ThemeProps } from '@t3n/theme';
 
 import { Box } from '../../Box';
 import { Text } from '../../Text';
-import { HeaderLink } from '../LegacyHeader';
+import { HeaderLink } from './LegacyHeaderLink';
 
 const ArrowDownIcon: React.FC = () => (
   <svg
