+++ packages/components/src/LegacyHeader/components/LegacyT3nNav.tsx
@@ -8,7 +8,7 @@ import { ThemeProps } from '@t3n/theme';
 import { Box } from '../../Box';
 import { LegacyUserMenu, LegacyUserMenuProps } from '../../LegacyUserMenu';
 import { Text } from '../../Text';
-import { HeaderLink } from '../LegacyHeader';
+import { HeaderLink } from './LegacyHeaderLink';
 import { MainNavDropdown } from './LegacyMainNav';
 
 const MainNavItem = styled(Box)<{ isSticky?: boolean }>`
