+++ packages/storybook/.storybook/main.js
@@ -41,10 +41,10 @@ module.exports = {
     '../src/stories/components/content/card.stories',
     '../src/stories/components/content/carousel.stories',
     '../src/stories/components/content/confirmDialog.stories',
-    '../src/stories/components/content/fastlyImage.stories',
     '../src/stories/components/content/heroArticleCard.stories',
     '../src/stories/components/content/icon.stories',
     '../src/stories/components/content/image.stories',
+    '../src/stories/components/content/imgixImage.stories',
     '../src/stories/components/content/jobCard.stories',
     '../src/stories/components/content/loader.stories',
     '../src/stories/components/content/modal.stories',
