+++ packages/storybook/src/stories/components/content/image.stories.tsx
@@ -1,7 +1,12 @@
-import React from 'react';
-import { boolean, number, text } from '@storybook/addon-knobs';
+import React, { useState } from 'react';
 
-import { Box, Grid, GridItem, H3, Image, Text } from '@t3n/components';
+// import { boolean, number, text } from '@storybook/addon-knobs';
+import {
+  Box,
+  Image,
+  // Grid, GridItem, H3,
+  Text,
+} from '@t3n/components';
 
 import { storyContainerDecorator } from '../../../utils/decorators';
 
@@ -11,136 +16,152 @@ export default {
   decorators: [storyContainerDecorator],
 };
 
-const options = {
-  range: true,
-  min: 0,
-  max: 100,
-  step: 5,
+const exampleImageSrc =
+  'https://images.t3n.de/news/wp-content/uploads/2021/11/shutterstock-1453899434-Internationale-Raumstation-ISS.jpg';
+
+const getFastlyClassFormSrc = (src: string) => {
+  // eslint-disable-next-line @typescript-eslint/no-unused-vars
+  const [_, fastlyClass] = src.split('?class=');
+
+  return fastlyClass;
 };
 
+const CurrentFastlyClassText = ({ src }: { src: string }) => (
+  <Text>
+    Currently used image optimization class:{' '}
+    <strong>
+      <code>{getFastlyClassFormSrc(src)}</code>
+    </strong>
+  </Text>
+);
+
+const ResponsiveImageHint = ({ src }: { src: string }) => (
+  <Box>
+    <Text>
+      Resize the window from smallest to largest to see the best otimized image
+      src being used depending on viewport width
+    </Text>
+    <CurrentFastlyClassText src={src} />
+    <Text>
+      💡{' '}
+      <em>
+        Hint: if you always see the same image class being used, reload the page
+        with cache disabled, because the largest option might be cached already
+      </em>
+    </Text>
+  </Box>
+);
+
+// All props customizable
 export const defaultStory = () => (
   <Image
-    src={text(
-      'Bild-URL',
-      'https://assets.t3n.sc/news/wp-content/uploads/2019/08/tier-e-scooter-hero.jpg'
-    )}
-    alt={text('Alt-Text', '')}
-    width={400}
-    height={200}
-    processConfiguration={{
-      fit: boolean('Crop?', true) ? 'crop' : undefined,
-      quality: number('Qualität', 65, options),
-      aspectRatio: '16:9',
-      crop: 'faces',
-    }}
+    src={exampleImageSrc}
+    width={1100}
+    height={685}
+    sizes="calc(80vw - 4rem)"
+    displayWidth="80%"
+    displayHeight="auto"
   />
 );
 
-defaultStory.storyName = 'Default';
+// Simple image without responsive sizes
+export const SimpleImage = () => (
+  <Image src={exampleImageSrc} width={800} height={450} />
+);
 
-export const CropModes = () => {
-  const directions = ['top', 'right', 'bottom', 'left'];
+// External image
+export const ExternalImage = () => (
+  <Image
+    src="https://images.pexels.com/photos/2253275/pexels-photo-2253275.jpeg?auto=compress&cs=tinysrgb&dpr=2&h=750&w=1260"
+    width={840}
+    height={560}
+  />
+);
+
+// Simple image without responsive sizes
+export const OptimizationClass = () => (
+  <Image
+    src={exampleImageSrc}
+    optimizationClass="hero-wide"
+    width={800}
+    height={367}
+  />
+);
+
+// Lazy demonstration
+export const Lazy = () => {
+  const [loadedSrc, setLoadedSrc] = useState('');
 
   return (
-    <>
-      <H3>Crop Mode - direktional</H3>
-      <Grid>
-        <GridItem width={[1, 1, 1 / 2]}>
-          <Text bold>Original</Text>
-          <Image
-            alt="Crop directional"
-            src="https://images.unsplash.com/photo-1492158244976-29b84ba93025"
-            width={400}
-          />
-        </GridItem>
-        <GridItem width={[1, 1 / 2]}>
-          <Grid>
-            {directions.map((direction) => (
-              <GridItem key={direction} width={[1 / 2]}>
-                <Text bold>{direction}</Text>
-                <Image
-                  alt="Crop faces example"
-                  src="https://images.unsplash.com/photo-1492158244976-29b84ba93025"
-                  width={100}
-                  height={100}
-                  processConfiguration={{
-                    fit: 'crop',
-                    crop: direction,
-                  }}
-                />
-              </GridItem>
-            ))}
-          </Grid>
-        </GridItem>
-      </Grid>
-
-      <H3>Crop mode auf Gesichter</H3>
-      <Grid>
-        <GridItem width={1 / 2}>
-          <Text bold>Original</Text>
-          <Image
-            alt="Crop faces example"
-            src="https://images.unsplash.com/photo-1521119989659-a83eee488004"
-            width={200}
-          />
-        </GridItem>
-        <GridItem width={1 / 2}>
-          <Text bold>Crop auf das Gesicht</Text>
-          <Image
-            alt="Crop faces example"
-            src="https://images.unsplash.com/photo-1521119989659-a83eee488004"
-            width={200}
-            height={200}
-            processConfiguration={{
-              crop: 'faces',
-              fit: 'facearea',
-              facepad: 2.0,
-            }}
-          />
-        </GridItem>
-      </Grid>
-    </>
+    <Box
+      height="200vh"
+      display="flex"
+      justifyContent="flex-start"
+      alignItems="flex-end"
+      position="relative"
+    >
+      <Box position="fixed" top="2rem" left="2rem">
+        <Text>
+          Scroll down to see the image. Once the image passes the viewport
+          threshold, the full image will be loaded instead of the blur
+          placeholder.
+        </Text>
+        <CurrentFastlyClassText src={loadedSrc} />
+      </Box>
+      <Image
+        src={exampleImageSrc}
+        width={1100}
+        height={685}
+        sizes={['calc(100vw - 4rem)', 'calc(80vw - 4rem)', 'calc(80vw - 4rem)']}
+        displayWidth={['100%', '80%', '80%']}
+        displayHeight="auto"
+        onLoad={(e) => setLoadedSrc(e.currentTarget.currentSrc)}
+      />
+    </Box>
   );
 };
 
-export const QualityConfiguration = () => {
-  const q = [10, 20, 30, 40, 50, 60, 70, 80, 90, 100];
-
-  return (
-    <Grid>
-      {q.map((el) => (
-        <GridItem width={[1 / 3]}>
-          <Text>Qualität: {el}</Text>
-          <Image
-            alt="Crop faces example"
-            src="https://images.unsplash.com/photo-1532910404247-7ee9488d7292"
-            width={300}
-            disableSrcSet
-            processConfiguration={{
-              quality: el,
-            }}
-          />
-        </GridItem>
-      ))}
-    </Grid>
-  );
-};
-
+// Responsive image
 export const Responsive = () => {
+  const [loadedSrc, setLoadedSrc] = useState('');
+
   return (
     <Box>
+      <ResponsiveImageHint src={loadedSrc} />
       <Image
-        src="https://images.unsplash.com/photo-1575315599174-f0c6d26625eb"
-        sizes="50vw"
-        width={[1, 1, 1 / 2, 1 / 3]}
-        alt="Responsive Image"
+        src={exampleImageSrc}
+        width={1100}
+        height={685}
+        sizes={['calc(100vw - 4rem)', 'calc(80vw - 4rem)', 'calc(80vw - 4rem)']}
+        displayWidth={['100%', '80%', '80%']}
+        displayHeight="auto"
+        onLoad={(e) => setLoadedSrc(e.currentTarget.currentSrc)}
       />
-      <Text small secondary>
-        Open the &quot;Network&quot; tab in your browser and check the
-        downloaded image file size. Now resize the window and check again. The
-        browser picks the appropriate size for the viewport and fetches the
-        corresponding image.
-      </Text>
+    </Box>
+  );
+};
+
+// Responsive image with custom class mapping
+export const CustomClassMapping = () => {
+  const [loadedSrc, setLoadedSrc] = useState('');
+
+  return (
+    <Box>
+      <ResponsiveImageHint src={loadedSrc} />
+      <Image
+        src={exampleImageSrc}
+        width={1200}
+        height={675}
+        sizes={['calc(100vw - 4rem)', 'calc(80vw - 4rem)', 'calc(80vw - 4rem)']}
+        classMapping={{
+          '630': 'hero-small',
+          '930': 'hero',
+          '1470': 'hero-large',
+        }}
+        displayWidth={['100%', '80%', '80%']}
+        displayHeight="auto"
+        onLoad={(e) => setLoadedSrc(e.currentTarget.currentSrc)}
+      />
     </Box>
   );
 };
